+++ app/build.gradle
@@ -1,0 +1,29 @@
+plugins {
+    id 'java'
+    id 'org.springframework.boot' version '3.1.3'
+    id 'io.spring.dependency-management' version '1.1.3'}
+
+group 'org.crypto'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.0.4'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.json:json:20230618'
+    implementation 'org.java-websocket:Java-WebSocket:1.5.4'
+
+    runtimeOnly 'org.postgresql:postgresql:42.5.4'
+    implementation 'com.h2database:h2:2.2.224'
+    testImplementation 'com.github.tomakehurst:wiremock:3.0.1'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'org.springframework.integration:spring-integration-test'
+
+}
+
+test {
+    useJUnitPlatform()
+}
