+++ app/src/main/java/com/crypto/domain/model/Crypto.java
