+++ app/src/main/java/com/crypto/infrastructure/adapters/inbound/dto/CryptoDto.java
