+++ app/src/main/java/com/crypto/infrastructure/adapters/outbound/http/BinanceApiDto.java
