+++ app/src/main/java/com/crypto/infrastructure/adapters/outbound/repository/entity/EntityMapper.java
