+++ app/src/main/resources/application.properties
@@ -1,0 +1,17 @@
+api_host=wss://stream.binance.com:9443/ws/
+#api_host=wss://stream.binance.com:9443/ws/btcusdt@trade
+spring.scan.package=com.rate.app
+
+
+spring.datasource.url=jdbc:h2:mem:crypto
+spring.datasource.username=postgres
+#spring.datasource.password=12345
+
+spring.jpa.properties.hibernate.default_schema=public
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.defer-datasource-initialization=always
+spring.sql.init.mode=always
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.show-sql=true
+spring.jpa.open-in-view=false
