+++ app/src/main/resources/schema.sql
@@ -1,0 +1,12 @@
+CREATE TABLE IF NOT EXISTS rates
+(
+    symbol VARCHAR(20) NOT NULL,
+    price DOUBLE PRECISION,
+    PRIMARY KEY (symbol)
+);
+
+CREATE TABLE IF NOT EXISTS CRYPTO_SYMBOLS
+(
+    symbol VARCHAR(20) NOT NULL,
+    PRIMARY KEY (symbol)
+);
