+++ app/src/main/resources/static/websocket.js
