+++ app/src/test/java/com/crypto/application/CreateSymbolActionTest.java
@@ -1,0 +1,8 @@
+package com.crypto.application;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+// TODO prosze napisac test
+class CreateSymbolActionTest {
+
+}
