+++ app/src/test/java/com/crypto/application/RateQueryTest.java
@@ -1,0 +1,8 @@
+package com.crypto.application;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+// TODO Prosze napisac test
+class RateQueryTest {
+
+}
