+++ app/src/test/java/com/crypto/domain/service/SymbolServiceTest.java
