+++ app/src/test/resources/__files.json/ethusdt-response.json
