+++ app/src/test/resources/__files.json/eurusdt-response.json
