+++ app/src/test/resources/__files.json/xrpusdt-response.json
