+++ build.gradle
@@ -1,54 +1,14 @@
 plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.1.3'
-    id 'io.spring.dependency-management' version '1.1.3'
+
 }
 
 group 'com.crypto'
 version '0.0.1-SNAPSHOT'
 
-java {
-    sourceCompatibility = '17'
-}
+
 repositories {
     mavenCentral()
 }
 
 dependencies {
-
-    //implementation 'org.springframework.boot:spring-boot-starter:3.1.3'
-    //implementation 'org.springframework.boot:spring-boot-starter-web:3.1.3'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.0.4'
-
-    //implementation 'org.springframework.boot:spring-boot-starter-integration'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    //implementation 'org.springframework.boot:spring-boot-starter-websocket'
-    //implementation 'org.springframework.integration:spring-integration-http'
-    //implementation 'org.springframework.integration:spring-integration-stomp'
-    //implementation 'org.springframework.integration:spring-integration-websocket'
-    implementation 'org.json:json:20230618'
-    implementation 'org.java-websocket:Java-WebSocket:1.5.4'
-
-    runtimeOnly 'org.postgresql:postgresql:42.5.4'
-
-    //testImplementation "com.github.tomakehurst:wiremock-jre8:2.35.0"
-    testImplementation 'com.github.tomakehurst:wiremock:3.0.1'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.integration:spring-integration-test'
-}
-
-    /*implementation 'org.java-websocket:Java-WebSocket:1.5.4'
-
-
-
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testImplementation 'org.mockito:mockito-junit-jupiter'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    testImplementation 'org.assertj:assertj-core:3.24.2'*/
-
-
-
-test {
-    useJUnitPlatform()
 }
